Refactor TopUpButton imports to consolidate path structure across components
