+++ apps/web/src/components/Settings/Funds/Balances.tsx
@@ -1,4 +1,4 @@
-import TopUpButton from "@/components/Shared/Account/Fund/TopUp/Button";
+import TopUpButton from "@/components/Shared/Account/TopUp/Button";
 import Loader from "@/components/Shared/Loader";
 import { ErrorMessage, Image } from "@/components/Shared/UI";
 import getTokenImage from "@/helpers/getTokenImage";
