+++ apps/web/src/components/Shared/GlobalModals.tsx
@@ -10,9 +10,9 @@ import { useReportPostModalStore } from "@/store/non-persisted/modal/useReportPo
 import { useSuperFollowModalStore } from "@/store/non-persisted/modal/useSuperFollowModalStore";
 import { useSuperJoinModalStore } from "@/store/non-persisted/modal/useSuperJoinModalStore";
 import { useSwitchAccountModalStore } from "@/store/non-persisted/modal/useSwitchAccountModalStore";
-import TopUp from "./Account/Fund/TopUp";
 import SuperFollow from "./Account/SuperFollow";
 import SwitchAccounts from "./Account/SwitchAccounts";
+import TopUp from "./Account/TopUp";
 import Auth from "./Auth";
 import { useSignupStore } from "./Auth/Signup";
 import SuperJoin from "./Group/SuperJoin";
